+++ crates/icon/src/lib.rs
@@ -11,6 +11,9 @@ pub const DEFAULT_ICON: char = '';
 pub const FOLDER_ICON: char = '';
 pub const DEFAULT_FILER_ICON: char = '';
 
+// Each added icon length is 4 bytes.
+pub const ICON_LEN: usize = 4;
+
 /// The type used to represent icons.
 ///
 /// This could be changed into different type later,