+++ crates/maple_cli/src/cmd/filter/mod.rs
@@ -8,7 +8,7 @@ use std::path::Path;
 use anyhow::Result;
 use fuzzy_filter::{fuzzy_filter_and_rank, truncate_long_matched_lines, Algo, Source};
 
-use icon::prepend_icon;
+use icon::{prepend_icon, ICON_LEN};
 
 /// Returns the info of the truncated top items ranked by the filtering score.
 fn process_top_items<T>(
@@ -16,14 +16,15 @@ fn process_top_items<T>(
     top_list: impl IntoIterator<Item = (String, T, Vec<usize>)>,
     winwidth: usize,
     enable_icon: bool,
+    add_icon: impl Fn(&str) -> String,
 ) -> (Vec<String>, Vec<Vec<usize>>, HashMap<String, String>) {
     let (truncated_lines, truncated_map) = truncate_long_matched_lines(top_list, winwidth, None);
     let mut lines = Vec::with_capacity(top_size);
     let mut indices = Vec::with_capacity(top_size);
     if enable_icon {
         for (text, _, idxs) in truncated_lines {
-            lines.push(prepend_icon(&text));
-            indices.push(idxs);
+            lines.push(add_icon(&text));
+            indices.push(idxs.into_iter().map(|x| x + ICON_LEN).collect());
         }
     } else {
         for (text, _, idxs) in truncated_lines {
@@ -51,6 +52,7 @@ pub fn run<I: Iterator<Item = String>>(
             ranked.into_iter().take(number),
             winwidth.unwrap_or(62),
             enable_icon,
+            prepend_icon,
         );
         if truncated_map.is_empty() {
             println_json!(total, lines, indices);
@@ -85,5 +87,6 @@ pub fn blines(
         number,
         false,
         winwidth,
+        false,
     )
 }