+++ crates/maple_cli/src/lib.rs
@@ -7,6 +7,20 @@ macro_rules! println_json {
   }
 }
 
+/// Combine json and println macro.
+///
+/// Neovim needs Content-length info when using stdio-based communication.
+macro_rules! print_json_with_length {
+  ( $( $field:expr ),+ ) => {
+    {
+      let msg = serde_json::json!({ $(stringify!($field): $field,)* });
+      if let Ok(s) = serde_json::to_string(&msg) {
+          println!("Content-length: {}\n\n{}", s.len(), s);
+      }
+    }
+  }
+}
+
 pub mod cmd;
 pub use {
     anyhow::Result,