+++ src/main.rs
@@ -25,9 +25,17 @@ fn run(maple: Maple) -> Result<()> {
         Cmd::Version => {
             version();
         }
+        Cmd::Helptags { meta_info } => maple_cli::cmd::helptags::run(meta_info)?,
         Cmd::RPC => {
             maple_cli::cmd::rpc::run_forever(std::io::BufReader::new(std::io::stdin()));
         }
+        Cmd::Blines { query, input } => {
+            maple_cli::cmd::filter::blines(&query, &input, maple.number, maple.winwidth)?;
+        }
+        Cmd::RipgrepForerunner { cmd_dir } => {
+            maple_cli::cmd::grep::run_forerunner(cmd_dir, maple.number, maple.enable_icon)?
+        }
+        Cmd::Cache { list } => maple_cli::cmd::cache::run(list)?,
         Cmd::Filter {
             query,
             input,
@@ -64,11 +72,9 @@ fn run(maple: Maple) -> Result<()> {
                     maple.number,
                     maple.enable_icon,
                     maple.winwidth,
+                    false,
                 )?;
             }
-        }
-        Cmd::Blines { query, input } => {
-            maple_cli::cmd::filter::blines(&query, &input, maple.number, maple.winwidth)?;
         }
         Cmd::Exec {
             cmd,
@@ -105,10 +111,6 @@ fn run(maple: Maple) -> Result<()> {
                 maple.number,
                 maple.enable_icon,
             )?;
-        }
-        Cmd::Helptags { meta_info } => maple_cli::cmd::helptags::run(meta_info)?,
-        Cmd::RipgrepForerunner { cmd_dir } => {
-            maple_cli::cmd::grep::run_forerunner(cmd_dir, maple.number, maple.enable_icon)?
         }
     }
     Ok(())